python/requirements.txt
# Core Flask application
Flask==3.0.0
requests==2.31.0
pandas==2.1.3
flask-cors==4.0.0

# RAG and Vector Database dependencies
python-dotenv==1.0.0
chromadb>=0.5.0
sentence-transformers>=2.3.0
torch>=2.2.0
transformers>=4.36.2

# Scheduling and automation
apscheduler>=3.10.0

# AI/ML
google-generativeai>=0.3.0
